moved EditorMediaListener implementation to ViewModel
